requirements.txt
langchain
python-dotenv==1.0.1
langchain-groq
